move fetch functions to definitions
