+++ core/base/src/constants/index.ts
@@ -15,3 +15,4 @@ export * as circle from "./circle.js";
 export * as contracts from "./contracts/index.js";
 export * as guardians from "./guardians.js";
 export * as tbtc from "./tbtc.js";
+export * as executor from "./executor.js";
